fix(model): remove legacy reference

+++ src/model.py
@@ -21,7 +21,7 @@ def generic_feed_forward(block_input: mtf.Tensor,
     intermediate_dimensions = [mtf.Dimension('_' + dim.name, dim.size) for dim in new_dimensions]
     with tf.variable_scope(f'feed_forward_{random.getrandbits(64):x}'):
         with tf.variable_scope('layer0'):
-            weight0 = mtf.get_variable(block_input.mesh, "weight0", reduced_dims + intermediate_dimensions, initializer=tf.orthogonal_initializer(), dtype=variable_dtype)
+            weight0 = mtf.get_variable(block_input.mesh, "weight0", reduced_dims + intermediate_dimensions, initializer=tf.orthogonal_initializer(), dtype=dtype)
             block_input = mtf.einsum([block_input, weight0], block_input.shape - reduced_dims + intermediate_dimensions)
         if dropout_rate > 0:
             block_input = mtf.dropout(block_input, 1 - dropout_rate)
